+++ Compare_Latest/Frameworks/AVFoundation/AVCompositionTrackReaderInspector.h
@@ -16,7 +16,7 @@
 @property(readonly, nonatomic, getter=_mutableComposition) struct OpaqueFigMutableComposition *mutableComposition;
 - (void)finalize;
 - (void)dealloc;
-- (id)_initWithAsset:(id)arg1 trackID:(int)arg2 trackIndex:(long)arg3;
+- (id)_initWithAsset:(id)arg1 trackID:(int)arg2 trackIndex:(long long)arg3;
 
 @end
