+++ Compare_Latest/Frameworks/AVFoundation/AVPlayerItemErrorLog.h
@@ -17,7 +17,7 @@
 
 - (id)_errorLogArray;
 @property(readonly, nonatomic) NSArray *events;
-- (unsigned int)extendedLogDataStringEncoding;
+- (unsigned long long)extendedLogDataStringEncoding;
 - (id)extendedLogData;
 - (void)dealloc;
 - (void)finalize;
