+++ Compare_Latest/Frameworks/AVFoundation/AVPlayerItemErrorLogEvent.h
@@ -17,7 +17,7 @@
 
 @property(readonly, nonatomic) NSString *errorComment;
 @property(readonly, nonatomic) NSString *errorDomain;
-@property(readonly, nonatomic) int errorStatusCode;
+@property(readonly, nonatomic) long long errorStatusCode;
 @property(readonly, nonatomic) NSString *playbackSessionID;
 @property(readonly, nonatomic) NSString *serverAddress;
 @property(readonly, nonatomic) NSString *URI;
