+++ Compare_Latest/Frameworks/AddressBook/CDStructures.h
@@ -14,8 +14,8 @@ struct CPSqliteStatement {
 };
 
 struct _NSRange {
-    unsigned int _field1;
-    unsigned int _field2;
+    unsigned long long _field1;
+    unsigned long long _field2;
 };
 
 struct sqlite3_stmt;
