+++ Compare_Latest/Frameworks/AddressBook/NSString-ABVCardAdditions.h
@@ -10,7 +10,7 @@
 - (id)abVCardKoshify;
 - (id)abStringByStrippingControlCharacters;
 - (id)abStringByRemovingCharactersFromSet:(id)arg1;
-- (id)abEscapeStringForUnichar:(unsigned short)arg1 and:(unsigned short)arg2 advance:(int *)arg3;
+- (id)abEscapeStringForUnichar:(unsigned short)arg1 and:(unsigned short)arg2 advance:(long long *)arg3;
 - (id)copyABVCardDataRepresentation;
 @end
