+++ Compare_Latest/Frameworks/AddressBookUI/ABActionSheetDelegate.h
@@ -14,7 +14,7 @@
     SEL _action;
 }
 
-- (void)actionSheet:(id)arg1 didDismissWithButtonIndex:(int)arg2;
+- (void)actionSheet:(id)arg1 didDismissWithButtonIndex:(long long)arg2;
 - (id)initWithActionSheetTarget:(id)arg1 action:(SEL)arg2;
 
 @end
