+++ Compare_Latest/Frameworks/AddressBookUI/ABContactViewDelegate-Protocol.h
@@ -10,8 +10,8 @@
 @protocol ABContactViewDelegate <NSObject, UITableViewDelegate>
 
 @optional
-- (void)contactView:(id)arg1 didSelectItemAtIndex:(int)arg2 inGroup:(id)arg3;
-- (float)contactView:(id)arg1 heightForItemAtIndex:(int)arg2 inGroup:(id)arg3;
-- (int)numberOfGroupsInContactView:(id)arg1;
+- (void)contactView:(id)arg1 didSelectItemAtIndex:(long long)arg2 inGroup:(id)arg3;
+- (double)contactView:(id)arg1 heightForItemAtIndex:(long long)arg2 inGroup:(id)arg3;
+- (long long)numberOfGroupsInContactView:(id)arg1;
 @end
