+++ Compare_Latest/Frameworks/AssetsLibrary/CDStructures.h
@@ -7,8 +7,8 @@
 #pragma mark Named Structures
 
 struct CGSize {
-    float _field1;
-    float _field2;
+    double _field1;
+    double _field2;
 };
 
 struct NSObject {
