+++ Compare_Latest/Frameworks/CoreData/PFUbiquityGlobalObjectIDCache.h
@@ -8,6 +8,7 @@
 
 @class NSMutableDictionary, NSRecursiveLock, NSString;
 
+// Not exported
 @interface PFUbiquityGlobalObjectIDCache : NSObject
 {
     NSString *_storeName;
@@ -21,7 +22,7 @@
 @property(retain, nonatomic) NSString *storeName; // @synthesize storeName=_storeName;
 - (void)purgeCache;
 - (id)createGlobalIDForPrimaryKeyString:(id)arg1 entityName:(id)arg2 andOwningPeerID:(id)arg3;
-- (id)createGlobalIDForPrimarKey:(unsigned int)arg1 entityName:(id)arg2 andOwningPeerID:(id)arg3;
+- (id)createGlobalIDForPrimarKey:(unsigned long long)arg1 entityName:(id)arg2 andOwningPeerID:(id)arg3;
 - (id)createGlobalIDForGlobalIDString:(id)arg1;
 - (id)createGlobalIDForCompressedString:(id)arg1 withEntityNames:(id)arg2 primaryKeys:(id)arg3 peerIDs:(id)arg4;
 - (void)dealloc;
