+++ Compare_Latest/Frameworks/CoreData/PFZipCentralDirectoryFileHeader.h
@@ -8,6 +8,7 @@
 
 @class NSData, NSString;
 
+// Not exported
 @interface PFZipCentralDirectoryFileHeader : NSObject
 {
     unsigned short _versionMadeBy;
@@ -51,8 +52,8 @@
 @property(nonatomic) unsigned short versionNeededToExtract; // @synthesize versionNeededToExtract=_versionNeededToExtract;
 @property(nonatomic) unsigned short versionMadeBy; // @synthesize versionMadeBy=_versionMadeBy;
 - (void)appendToData:(id)arg1;
-- (unsigned int)loadFromBytes:(const char *)arg1 offset:(unsigned int)arg2;
-- (unsigned int)loadFromData:(id)arg1 offset:(unsigned int)arg2;
+- (unsigned long long)loadFromBytes:(const char *)arg1 offset:(unsigned long long)arg2;
+- (unsigned long long)loadFromData:(id)arg1 offset:(unsigned long long)arg2;
 - (id)description;
 - (void)dealloc;
 - (id)init;
